applications/ml-training/requirements.txt
# Core ML libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Model libraries
xgboost==2.0.0
torch==2.1.0

# MLflow
mlflow==2.8.1
boto3==1.28.85

# Spark (for feature loading)
pyspark==3.5.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
pyyaml==6.0.1
joblib==1.3.2
tqdm==4.66.1

# AWS/S3
s3fs==2023.10.0